lots of stuff

+++ js/signup.js
@@ -10,12 +10,23 @@ function registerUser() {
   const lastName = document.getElementById('lastName').value;
   const email = document.getElementById('emailAddress').value;
   const password = document.getElementById('password').value;
-  
-  window.localStorage.setItem('firstName', firstName)
-  window.localStorage.setItem('lastName', lastName)
-  window.localStorage.setItem('email', email)
-  
+  const confirmPassword = document.getElementById('confirmPassword').value;
+  if (password !== confirmPassword) {
+    alert('Password does not match');
+    return;
+  }
 
 
+  window.localStorage.setItem('firstName', firstName);
+  window.localStorage.setItem('lastName', lastName);
+  window.localStorage.setItem('email', email);
   
+  window.location.href = "../";
 }
+
+
+
+
+
+
+
